Configure for local testing
The github-pages configurations does not work if we do not have an SSL
certificate on local machine(which developers usually don't). Hence the
testing needs to be done in this environment.

To test some new feature, we first test it on the local branch, if all
is good then stash the changes and pop the stash onto master branch.
After committing on the master branch, we synchronize the local branch
with the master branch using rebase.